+++ griffon/services/core_queries.py
@@ -217,20 +217,13 @@ def async_retrieve_sources(self, purl):
     params = {
         "limit": 200,
         "root_components": "True",
-        "released_components": "True",
         "provides": purl,
-        "include_fields": "type,nvr,purl,name,namespace,download_url,related_url",
+        "include_fields": "type,nvr,purl,name,version,namespace,download_url,related_url",
     }
-    # TODO: remove cnt and max_result and cnt > 10000 on next stage-> prod deployment
-    cnt = self.components.count(**params)
-    if cnt == 0:
-        return []
     try:
-        if cnt > 10000:
-            return list(self.components.retrieve_list_iterator_async(max_results=5000, **params))
-        return list(self.components.retrieve_list_iterator_async(max_results=10000, **params))
+        return list(self.components.retrieve_list_iterator_async(**params))
     except Exception as e:
-        logger.warning(f"{type(e).__name__} - problem retrieving all of {purl} {cnt} sources.")
+        logger.warning(f"{type(e).__name__} - problem retrieving all of {purl} sources.")
         return []
 
 
@@ -239,27 +232,44 @@ def async_retrieve_upstreams(self, purl):
         "limit": 200,
         "root_components": "True",
         "upstreams": purl,
-        "include_fields": "type,nvr,purl,name,namespace,download_url,related_url",
+        "include_fields": "type,nvr,purl,name,version,namespace,download_url,related_url",
     }
-    # TODO: remove cnt and max_result and cnt > 10000 on next stage-> prod deployment
-    cnt = self.components.count(**params)
-    if cnt == 0:
-        return []
     try:
-        if cnt > 10000:
-            return list(self.components.retrieve_list_iterator_async(max_results=5000, **params))
-        return list(self.components.retrieve_list_iterator_async(max_results=10000, **params))
+        return list(self.components.retrieve_list_iterator_async(**params))
     except Exception as e:
-        logger.warning(f"{type(e).__name__} - problem retrieving all of {purl} {cnt} upstreams.")
+        logger.warning(f"{type(e).__name__} - problem retrieving all of {purl} upstreams.")
         return []
 
 
-def process_component(session, c):
+def async_retrieve_provides(self, urlparams, purl):
+    params = {
+        "limit": 200,
+        "sources": purl,
+        "include_fields": "type,arch,nvr,purl,version,name,namespace,download_url,related_url",
+    }
+    if "name" in urlparams:
+        params["name"] = urlparams["name"]
+    if "provides_name" in urlparams:
+        params["name"] = urlparams["provides_name"]
+    if "re_name" in urlparams:
+        params["re_name"] = urlparams["re_name"]
+    if "re_provides_name" in urlparams:
+        params["re_name"] = urlparams["re_provides_name"]
+    if "namespace" in urlparams:
+        params["namespace"] = urlparams["namespace"]
+    try:
+        return list(self.components.retrieve_list_iterator_async(**params))
+    except Exception as e:
+        logger.warning(f"{type(e).__name__} - problem retrieving all of {purl} provides.")
+        return []
+
+
+def process_component(session, urlparams, c):
     """perform any neccessary sub retrievals."""
-    # only process provided components
-    if c.software_build is None:
+    if c.sources:
         c.sources = async_retrieve_sources(session, c.purl)
-        c.upstreams = async_retrieve_upstreams(session, c.purl)
+    c.upstreams = async_retrieve_upstreams(session, c.purl)
+    c.provides = async_retrieve_provides(session, urlparams, c.purl)
     return c
 
 
@@ -350,8 +360,12 @@ class products_containing_component_query:
                 **search_latest_params
             )
             status.update(f"found {latest_components_cnt} latest root component(s).")  # noqa
-            for c in latest_components:
-                results.append(c)
+            with multiprocessing.Pool() as pool:
+                for processed_component in pool.map(
+                    partial(process_component, self.corgi_session, search_latest_params),
+                    latest_components,
+                ):
+                    results.append(processed_component)
 
             if not self.no_community:
                 status.update("searching latest community root component(s).")
@@ -367,10 +381,14 @@ class products_containing_component_query:
                     )
                 )
                 status.update(
-                    f"found {community_component_cnt} latest community root component(s), retrieving sources & upstreams."  # noqa
+                    f"found {community_component_cnt} latest community root component(s)- retrieving children, sources & upstreams."  # noqa
                 )
-                for c in latest_community_components:
-                    results.append(c)
+                with multiprocessing.Pool() as pool:
+                    for processed_component in pool.map(
+                        partial(process_component, self.corgi_session, search_latest_params),
+                        latest_community_components,
+                    ):
+                        results.append(processed_component)
 
         if self.search_provides:
             search_provides_params = copy.deepcopy(params)
@@ -390,10 +408,14 @@ class products_containing_component_query:
                 **search_provides_params
             )
             status.update(
-                f"found {latest_components_cnt} latest provides child component(s), retrieving sources & upstreams."  # noqa
+                f"found {latest_components_cnt} latest provides child component(s)- retrieving children, sources & upstreams."  # noqa
             )
-            for c in latest_components:
-                results.append(c)
+            with multiprocessing.Pool() as pool:
+                for processed_component in pool.map(
+                    partial(process_component, self.corgi_session, search_provides_params),
+                    latest_components,
+                ):
+                    results.append(processed_component)
 
             if not self.no_community:
                 status.update("searching latest community provided child component(s).")
@@ -409,10 +431,14 @@ class products_containing_component_query:
                     )
                 )
                 status.update(
-                    f"found {community_component_cnt} latest community provided child component(s), retrieving sources & upstreams."  # noqa
+                    f"found {community_component_cnt} latest community provided child component(s)- retrieving children, sources & upstreams."  # noqa
                 )
-                for c in latest_community_components:
-                    results.append(c)
+                with multiprocessing.Pool() as pool:
+                    for processed_component in pool.map(
+                        partial(process_component, self.corgi_session, search_provides_params),
+                        latest_community_components,
+                    ):
+                        results.append(processed_component)
 
         if self.search_upstreams:
             search_upstreams_params = copy.deepcopy(params)
@@ -435,10 +461,11 @@ class products_containing_component_query:
             )
             with multiprocessing.Pool() as pool:
                 status.update(
-                    f"found {latest_components_cnt} latest upstreams child component(s), retrieving sources & upstreams."  # noqa
+                    f"found {latest_components_cnt} latest upstreams child component(s)- retrieving children, sources & upstreams."  # noqa
                 )
                 for processed_component in pool.map(
-                    partial(process_component, self.corgi_session), latest_components
+                    partial(process_component, self.corgi_session, search_upstreams_params),
+                    latest_components,
                 ):
                     results.append(processed_component)
             if not self.no_community:
@@ -456,10 +483,10 @@ class products_containing_component_query:
                 )
                 with multiprocessing.Pool() as pool:
                     status.update(
-                        f"found {community_component_cnt} latest community provided child component(s), retrieving sources & upstreams."  # noqa
+                        f"found {community_component_cnt} latest community provided child component(s)- retrieving children, sources & upstreams."  # noqa
                     )
                     for processed_component in pool.map(
-                        partial(process_component, self.community_session),
+                        partial(process_component, self.community_session, search_upstreams_params),
                         latest_community_components,
                     ):
                         results.append(processed_component)
@@ -484,10 +511,11 @@ class products_containing_component_query:
             )
             with multiprocessing.Pool() as pool:
                 status.update(
-                    f"found {related_url_components_cnt} related url component(s), retrieving sources & upstreams."  # noqa
+                    f"found {related_url_components_cnt} related url component(s)- retrieving children, sources & upstreams."  # noqa
                 )
                 for processed_component in pool.map(
-                    partial(process_component, self.corgi_session), related_url_components
+                    partial(process_component, self.corgi_session, search_related_url_params),
+                    related_url_components,
                 ):
                     results.append(processed_component)
             if not self.no_community:
@@ -504,10 +532,12 @@ class products_containing_component_query:
                 )
                 with multiprocessing.Pool() as pool:
                     status.update(
-                        f"found {latest_community_url_components_cnt} related url community component(s), retrieving sources & upstreams."  # noqa
+                        f"found {latest_community_url_components_cnt} related url community component(s)- retrieving children, sources & upstreams."  # noqa
                     )
                     for processed_component in pool.map(
-                        partial(process_component, self.community_session),
+                        partial(
+                            process_component, self.community_session, search_related_url_params
+                        ),
                         latest_community_url_components,
                     ):
                         results.append(processed_component)
@@ -524,7 +554,7 @@ class products_containing_component_query:
                 search_all_params["namespace"] = self.ns
             if not (self.include_inactive_product_streams):
                 search_all_params["active_streams"] = "True"
-            search_all_params["released_components"] = "True"
+            # search_all_params["released_components"] = "True"
             all_components_cnt = self.corgi_session.components.count(**search_all_params)
             status.update(f"found {all_components_cnt} all component(s).")
             # TODO: remove max_results
@@ -534,10 +564,11 @@ class products_containing_component_query:
             status.update(f"found {all_components_cnt} all component(s).")
             with multiprocessing.Pool() as pool:
                 status.update(
-                    f"found {all_components_cnt} all component(s), retrieving sources & upstreams."  # noqa
+                    f"found {all_components_cnt} all component(s)- retrieving children, sources & upstreams."  # noqa
                 )
                 for processed_component in pool.map(
-                    partial(process_component, self.corgi_session), all_components
+                    partial(process_component, self.corgi_session, search_all_params),
+                    all_components,
                 ):
                     results.append(processed_component)
 
@@ -556,10 +587,10 @@ class products_containing_component_query:
                 )
                 with multiprocessing.Pool() as pool:
                     status.update(
-                        f"found {all_community_components_cnt} community all component(s), retrieving sources & upstreams."  # noqa
+                        f"found {all_community_components_cnt} community all component(s)- retrieving children, sources & upstreams."  # noqa
                     )
                     for processed_component in pool.map(
-                        partial(process_component, self.community_session),
+                        partial(process_component, self.community_session, search_all_params),
                         all_community_components,
                     ):
                         results.append(processed_component)
@@ -575,7 +606,7 @@ class products_containing_component_query:
                 search_all_roots_params["namespace"] = self.ns
             if not (self.include_inactive_product_streams):
                 search_all_roots_params["active_streams"] = "True"
-            search_all_roots_params["released_components"] = "True"
+            # search_all_roots_params["released_components"] = "True"
             all_src_components_cnt = self.corgi_session.components.count(**search_all_roots_params)
             status.update(f"found {all_src_components_cnt} all root component(s).")
             all_src_components = self.corgi_session.components.retrieve_list_iterator_async(
@@ -619,7 +650,8 @@ class products_containing_component_query:
             with multiprocessing.Pool() as pool:
                 status.update(f"found {upstream_components_cnt} upstream component(s).")
                 for processed_component in pool.map(
-                    partial(process_component, self.corgi_session), upstream_components
+                    partial(process_component, self.corgi_session, search_all_upstreams_params),
+                    upstream_components,
                 ):
                     results.append(processed_component)
             if not self.no_community:
@@ -636,10 +668,12 @@ class products_containing_component_query:
                 )
                 with multiprocessing.Pool() as pool:
                     status.update(
-                        f"found {commmunity_upstream_components_cnt} community upstream component(s), retrieving sources & upstreams."  # noqa
+                        f"found {commmunity_upstream_components_cnt} community upstream component(s)- retrieving children, sources & upstreams."  # noqa
                     )
                     for processed_component in pool.map(
-                        partial(process_component, self.community_session),
+                        partial(
+                            process_component, self.community_session, search_all_upstreams_params
+                        ),
                         commmunity_upstream_components,
                     ):
                         results.append(processed_component)
@@ -706,10 +740,11 @@ class products_containing_component_query:
             )
             with multiprocessing.Pool() as pool:
                 status.update(
-                    f"found {all_community_components_cnt} community all component(s), retrieving sources & upstreams."  # noqa
+                    f"found {all_community_components_cnt} community all component(s)- retrieving children, sources & upstreams."  # noqa
                 )
                 for processed_component in pool.map(
-                    partial(process_component, self.community_session), all_community_components
+                    partial(process_component, self.community_session, search_community_params),
+                    all_community_components,
                 ):
                     results.append(processed_component)